refactor(connect): separate Blockchain class
